Fix graphql syntax error
